\name{Autoplotprotein-package}
\alias{Autoplotprotein-package}
\docType{package}
\title{
\packageTitle{Autoplotprotein}
}
\description{
\packageDescription{Autoplotprotein}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Autoplotprotein}
\packageIndices{Autoplotprotein}

}
\author{
\packageAuthor{Autoplotprotein}

Maintainer: \packageMaintainer{Autoplotprotein}
}
\references{
https://cran.r-project.org/doc/manuals/R-exts.html}

\keyword{ package }
\seealso{
code{\link{help}}

}

